requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
anyio==3.6.2
apache-airflow==2.6.1
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.2
apispec==5.2.2
argcomplete==3.0.8
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
Babel==2.12.1
beautifulsoup4==4.12.2
blinker==1.6.2
boto3==1.26.135
botocore==1.29.135
cachelib==0.9.0
cattrs==22.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
configparser==5.3.0
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.3.0
croniter==1.3.14
cryptography==40.0.2
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.3.0
docutils==0.20
email-validator==1.3.1
exceptiongroup==1.1.1
Flask==2.2.5
Flask-AppBuilder==4.3.0
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.3.3
graphviz==0.20.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
lazy-object-proxy==1.9.0
limits==3.4.0
linkify-it-py==2.0.2
lockfile==0.12.2
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.5
mdurl==0.1.2
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.3
ordered-set==4.1.0
outcome==1.2.0
packaging==21.3
pandas==2.0.1
pathspec==0.9.0
pendulum==2.1.2
pluggy==1.0.0
prison==0.2.1
psutil==5.9.5
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
pyrsistent==0.19.3
pyscopg2==66.0.2
PySocks==1.7.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
redis==4.5.5
requests==2.30.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.3.5
rich_argparse==1.1.0
s3transfer==0.6.1
selenium==4.9.1
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spotipy==2.23.0
SQLAlchemy==1.4.48
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
wsproto==1.2.0
WTForms==3.0.1
yarl==1.9.2
